test/lib/test_utils: add boost_test_print_type() for ranges
since some types do not have operator<< anymore, and we will ditch the
homebrew generic operator<< formatters for STL container types. and
these formatters are mostly used by tests for helpers like
`BOOST_REQUIRE_EQUAL`, so let's provide a replacement for these
generic formatters using {fmt}.

in this change, a formatter specifically provided for test is added,
since we have specializations for string-alike types, this formatter
does not apply to them. we could consolidate the string-alike types
latter on, but let's address this use case first.

Refs #13245

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>