storage_service: update system.cdc_local in topology_state_load
When the node with CDC enabled and with the topology on raft disabled,
bootstraps it reads system.cdc_local for the last generation.
Nodes with both enabled use group0 to get the last generation.

In the following scenario with a cluster of one node:
1. the node is created with CDC and the topology on raft enabled
2. the user creates table T
3. the node is restarted in the recovery mode
4. the CDC log of T is extended with new entries
5. the node restarts in normal mode
The newly created generation is seen in system_distributed.cdc_generation_timestamps
but not in system.cdc_generations_v3, thus there are used streams that the CDC
based on raft doesn't know about.

Save the last CDC generation in the system.cdc_local during loading the topology
state so that it is visible for CDC not based on raft.

Fixes #17819