sstables_loader: Fix load and stream for tablet
When the token range of a sstable is not completely within a tablet
token range, we need to stream all mutations in the sstables. The range
passed to the stream_sstables() should cover all the ranges in the
sstables_partially_contained list.

Otherwise, load and stream will not stream all the data within the
sstables.

While at it, two small optimizations are added.

- Call stream_sstables only when the list is not empty

- Skip the loop if all the sstables are processed

Fixes: #17733