compaction/compaction_manager: introduce perform_bloom_filter_regeneration()
Iterate over all sstables in the specified keyspace and tables and check
whether they should have their filtered regenerated according to the
provided conditions.
Currently, the only condition is filter_size_mismatch_tolerance.
Given the actual filter size, and the calculated filter size (using the
actual partition count and the current fp chance), this condition for
regenerating the filter is fulfilled if:

    actual < calculated / tolerance || actual > calculated * tolerance

If no conditions are provided, all sstable filters are regenerated.