github: sync-labels: use more descriptive name for workflow
"label-sync" is not very helpful for developers to understand what
this workflow is for.

the "name" field of a job shows in the webpage on github of the
pull request against which the job is performed, so if the author
or reviewer checks the status of the pull request, he/she would
notice these names aside of the workflow's name. for this very
job, what we have now is:

```
Sync labels / label-sync
```

after this change it will be:
```
Sync labels / Synchronize labels between PR and the issue(s) fixed by it
```

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>