db/hints: Provide a way to stop resource manager from scanning
We introduce an API for making resource manager
suspend and resume scanning the hint directory.
It will be necessary in an upcoming commit
when we will be renaming hint directories
and possibly remove some of them. Race conditions
are unacceptable, so preventing resource manager
from accessing the directory will be crucial.