.github: skip "raison" when running codespell workflow
codespell workflow checks for misspellings to identify common
mispellings. it considers "raison" in "raison d'etre" (the accent
mark over "e" is removed , so the commit message can be encoded in
ASCII), to the misspelling of "reason" or "raisin". apparently, the
dictionary it uses does not include les mots francais les plus
utilises.

so, in this change, let's ignore "raison" for this very use case,
before we start the l10n support of the document.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>