schema_tables: merge_schema: use freeze_gently
When applying mutations on a different shard
we're freezing the mutations for transferring
them to the other shard and there, we're unfreezing them.

In commit 731b5c51204c5b2fba8344804f91da4cd318b03e,
unfreezing became async to prevent a reactor stall,
but the freezing part is still synchronous and
it may cause a reactor stall as well with a large
number of mutations.

Use freeze_gently to prevent this potential reactor
stall too.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>